requirements.txt
# AI & LLM (FREE)
groq==0.4.1
sentence-transformers==2.3.1
torch>=2.2.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
plotly==5.18.0
kaleido==0.2.1
